# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 관리의달인 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /credits/wallet:
    get:
      operationId: getCreditWallet
      tags: [commissions]
      summary: Get partner credit wallet (self for partner, target for admin)
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditWallet"
  /credits/ledger:
    get:
      operationId: listCreditLedger
      tags: [commissions]
      summary: List credit ledger entries
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditLedgerEntry"
  /credits/preview:
    get:
      operationId: previewCreditCost
      tags: [commissions]
      summary: Preview credit deduction for an RFQ
      parameters:
        - name: rfqId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCostPreview"
  /credits/adjust:
    post:
      operationId: adjustCredits
      tags: [commissions]
      summary: Manual credit/debit adjustment by HQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCreditsBody"
      responses:
        "201":
          description: Posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditLedgerEntry"
...
  /credits/category-pricing:
    get:
      operationId: listCreditCategoryPricing
      tags: [commissions]
      summary: List credit tier/cost per category
      responses:
        "200":
          description: Pricing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditCategoryPricing"
    put:
      operationId: upsertCreditCategoryPricing
      tags: [commissions]
      summary: Upsert credit pricing for a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCreditCategoryPricingBody"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCategoryPricing"
...
components:
  schemas:
...
    CreditWallet:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        balance:
          type: integer
        pointsBalance:
          type: integer
        creditsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - vendorId
        - balance
        - pointsBalance
    CreditLedgerEntry:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        amount:
          type: integer
        kind:
          type: string
          enum: [consumption, refund, manual_credit, manual_debit, package_purchase, rebate, adjustment, bonus_points]
        source:
          type: string
          enum: [manual, package_purchase, refund, rebate, consumption, adjustment, system]
        pointsAmount:
          type: integer
        rfqId:
          type: ["integer", "null"]
        quoteId:
          type: ["integer", "null"]
        relatedLedgerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        actorId:
          type: ["integer", "null"]
        actorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - vendorId
        - amount
        - kind
        - source
        - pointsAmount
        - createdAt
    CreditCostPreview:
      type: object
      properties:
        rfqId:
          type: integer
        baseCost:
          type: integer
        tier:
          type: integer
        largeBuildingMultiplier:
          type: number
        isPremium:
          type: boolean
        totalCost:
          type: integer
        reason:
          type: array
          items:
            type: string
        creditsEnabled:
          type: boolean
        isPremiumRfq:
          type: boolean
        slotLimit:
          type: ["integer", "null"]
        slotsRemaining:
          type: ["integer", "null"]
      required:
        - rfqId
        - baseCost
        - tier
        - largeBuildingMultiplier
        - isPremium
        - totalCost
        - reason
    AdjustCreditsBody:
      type: object
      properties:
        vendorId:
          type: integer
        amount:
          type: integer
        kind:
          type: string
          enum: [manual_credit, manual_debit, adjustment, package_purchase, rebate, bonus_points]
        pointsAmount:
          type: integer
        notes:
          type: string
      required:
        - vendorId
        - amount
        - kind
        - notes
...
    CreditCategoryPricing:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        tier:
          type: integer
        creditCost:
          type: integer
        description:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - category
        - tier
        - creditCost
    UpsertCreditCategoryPricingBody:
      type: object
      properties:
        category:
          type: string
        tier:
          type: integer
        creditCost:
          type: integer
        description:
          type: ["string", "null"]
      required:
        - category
        - tier
        - creditCost
...