# lib/api-zod/src/generated/types/creditLedgerEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 관리의달인 API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreditLedgerEntryKind } from "./creditLedgerEntryKind";
import type { CreditLedgerEntrySource } from "./creditLedgerEntrySource";

export interface CreditLedgerEntry {
  id: number;
  vendorId: number;
  amount: number;
  kind: CreditLedgerEntryKind;
  source: CreditLedgerEntrySource;
  pointsAmount: number;
  /** @nullable */
  rfqId?: number | null;
  /** @nullable */
  quoteId?: number | null;
  /** @nullable */
  relatedLedgerId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  actorId?: number | null;
  /** @nullable */
  actorName?: string | null;
  createdAt: Date;
}